fixes #113
