--- src/lib/3_Organism/Footer.svelte
@@ -1,2 +1,0 @@
-<footer>
-</footer>
