+++ src/lib/ReusableComponents/Footer.svelte
@@ -1,0 +1,2 @@
+<footer>
+</footer>
