+++ src/lib/ReusableComponents/Icon.svelte
@@ -1,0 +1,129 @@
+<script>
+  export let name = '';
+</script>
+
+{#if name == 'bedrijf'}
+  <svg width="17" height="18" viewBox="0 0 17 18" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M6 1.92773C3.08702 2.95733 1 5.73542 1 9.00098C1 13.1431 4.35787 16.501 8.5 16.501C12.6422 16.501 16 13.1431 16 9.00098C16 5.73542 13.913 2.95733 11 1.92773" stroke="#0275FF" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+  </svg>
+
+{:else if name == "contrast-errors"}
+  <svg width="43" height="44" viewBox="0 0 43 44" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M21.3873 41.8856C32.316 41.8856 41.1755 32.9338 41.1755 21.8913C41.1755 10.8487 32.316 1.89697 21.3873 1.89697C10.4586 1.89697 1.59912 10.8487 1.59912 21.8913C1.59912 32.9338 10.4586 41.8856 21.3873 41.8856Z" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5874 9.5542L5.99671 9.5542" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5874 34.2285L5.99671 34.2285" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5874 21.4673L1.59933 21.4673" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5874 15.937L2.79862 15.937" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5874 27.8457L2.79862 27.8457" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.3868 41.8851C32.3155 41.8851 41.175 32.9334 41.175 21.8908C41.175 10.8482 32.3155 1.89648 21.3868 1.89648C10.4581 1.89648 1.59863 10.8482 1.59863 21.8908C1.59863 32.9334 10.4581 41.8851 21.3868 41.8851Z" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5869 9.55371L5.99622 9.55371" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5869 34.228L5.99622 34.228" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5869 21.4668L1.59884 21.4668" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5869 15.9365L2.79813 15.9365" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M21.5869 27.8452L2.79813 27.8452" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+  </svg>
+
+{:else if name == "missing-alt-texts"}
+  <svg width="43" height="44" viewBox="0 0 43 44" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M4.78663 41.9884H38.2742C39.9555 41.9884 41.3185 40.6113 41.3185 38.9124V5.07603C41.3185 3.37719 39.9555 2 38.2742 2H4.78663C3.1053 2 1.74231 3.37719 1.74231 5.07603V38.9124C1.74231 40.6113 3.1053 41.9884 4.78663 41.9884Z" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M5.5481 41.9887L27.7716 20.8255C28.0451 20.5953 28.3899 20.4692 28.7458 20.4692C29.1018 20.4692 29.4465 20.5953 29.72 20.8255L41.3189 29.223" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M13.9204 18.919C16.4424 18.919 18.4868 16.8532 18.4868 14.305C18.4868 11.7567 16.4424 9.69092 13.9204 9.69092C11.3984 9.69092 9.35388 11.7567 9.35388 14.305C9.35388 16.8532 11.3984 18.919 13.9204 18.919Z" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+  </svg>
+
+{:else if name == "structural-errors"}
+  <svg width="48" height="46" viewBox="0 0 48 46" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M6.90198 33.1532V26.2301C6.90198 25.312 7.26293 24.4316 7.90538 23.7824C8.54786 23.1333 9.41923 22.7686 10.3278 22.7686H37.7347C38.6432 22.7686 39.5147 23.1333 40.1571 23.7824C40.7994 24.4316 41.1605 25.312 41.1605 26.2301V33.1532" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M24.031 12.3853V33.1545" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M18.892 5.46154V8.92308C18.892 10.8348 20.4258 12.3846 22.3178 12.3846H25.7437C27.6357 12.3846 29.1695 10.8348 29.1695 8.92308V5.46154C29.1695 3.5498 27.6357 2 25.7437 2H22.3178C20.4258 2 18.892 3.5498 18.892 5.46154Z" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M18.892 36.6153V40.0769C18.892 41.9887 20.4258 43.5384 22.3178 43.5384H25.7437C27.6357 43.5384 29.1695 41.9887 29.1695 40.0769V36.6153C29.1695 34.7035 27.6357 33.1538 25.7437 33.1538H22.3178C20.4258 33.1538 18.892 34.7035 18.892 36.6153Z" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M1.76294 36.6153V40.0769C1.76294 41.9887 3.29675 43.5384 5.18879 43.5384H8.61465C10.5067 43.5384 12.0405 41.9887 12.0405 40.0769V36.6153C12.0405 34.7035 10.5067 33.1538 8.61465 33.1538H5.18879C3.29675 33.1538 1.76294 34.7035 1.76294 36.6153Z" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M36.0209 36.6153V40.0769C36.0209 41.9887 37.5546 43.5384 39.4467 43.5384H42.8726C44.7647 43.5384 46.2984 41.9887 46.2984 40.0769V36.6153C46.2984 34.7035 44.7647 33.1538 42.8726 33.1538H39.4467C37.5546 33.1538 36.0209 34.7035 36.0209 36.6153Z" stroke="#0275FF" stroke-width="3" stroke-linecap="round" stroke-linejoin="round"/>
+  </svg>
+
+{:else if name == "aria-label-errors"}
+  <svg width="44" height="31" viewBox="0 0 44 31" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M2 6.5C2 4.37868 2 3.31803 2.65825 2.659C3.31652 2 4.37596 2 6.49486 2H31.0584C32.236 2 32.8249 2 33.3218 2.26624C33.8187 2.53246 34.1452 3.02292 34.7983 4.00385L40.7916 13.0039C41.5973 14.2137 42 14.8187 42 15.5C42 16.1813 41.5973 16.7863 40.7916 17.9961L34.7983 26.9961C34.1452 27.9771 33.8187 28.4674 33.3218 28.7338C32.8249 29 32.236 29 31.0584 29H6.49486C4.37596 29 3.31652 29 2.65825 28.341C2 27.6819 2 26.6213 2 24.5V6.5Z" stroke="#0275FF" stroke-width="3"/>
+  </svg>
+
+{:else if name == "detected-errors"}
+  <svg width="45" height="44" viewBox="0 0 45 44" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M11.5528 11.0878L33.4462 33.8087M11.5528 33.8087L33.4462 11.0878M42.2877 6.99951V36.9881C42.2877 39.7496 40.0491 41.9881 37.2877 41.9881H7.7113C4.94988 41.9881 2.7113 39.7496 2.7113 36.9881V6.99951C2.7113 4.23809 4.94988 1.99951 7.7113 1.99951H37.2877C40.0491 1.99951 42.2877 4.23809 42.2877 6.99951Z" stroke="#0275FF" stroke-width="4" stroke-linecap="round" stroke-linejoin="round"/>
+  </svg>
+
+{:else if name == "arrow-left"}
+  <svg width="6" height="11" viewBox="0 0 6 11" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M5 1L2.20676 4.02601C1.51868 4.77143 1.49766 5.91393 2.15785 6.68416L5 10" stroke="#5D666A" stroke-width="2" stroke-linecap="round"/>
+  </svg>
+
+{:else if name == "arrow-right"}
+  <svg width="6" height="11" viewBox="0 0 6 11" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M1 1L3.79324 4.02601C4.48132 4.77143 4.50234 5.91393 3.84215 6.68416L1 10" stroke="#5D666A" stroke-width="2" stroke-linecap="round"/>
+  </svg>
+
+{:else if name == "dashboard"}
+  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="29" viewBox="0 0 16 16" fill="none">
+    <path d="M2.42307 15.0831C2.00748 15.0831 1.65251 14.936 1.35818 14.6416C1.06383 14.3473 0.916656 13.9923 0.916656 13.5767V2.42292C0.916656 2.00732 1.06383 1.65236 1.35818 1.35802C1.65251 1.06368 2.00748 0.916504 2.42307 0.916504H13.5769C13.9925 0.916504 14.3474 1.06368 14.6418 1.35802C14.9361 1.65236 15.0833 2.00732 15.0833 2.42292V13.5767C15.0833 13.9923 14.9361 14.3473 14.6418 14.6416C14.3474 14.936 13.9925 15.0831 13.5769 15.0831H2.42307ZM6.54166 13.8332V8.6248H2.16664V13.5767C2.16664 13.6515 2.19068 13.7129 2.23876 13.761C2.28684 13.8091 2.34828 13.8332 2.42307 13.8332H6.54166ZM7.79162 13.8332H13.5769C13.6517 13.8332 13.7131 13.8091 13.7612 13.761C13.8093 13.7129 13.8333 13.6515 13.8333 13.5767V8.6248H7.79162V13.8332ZM2.16664 7.37484H13.8333V2.42292C13.8333 2.34813 13.8093 2.28669 13.7612 2.23861C13.7131 2.19052 13.6517 2.16648 13.5769 2.16648H2.42307C2.34828 2.16648 2.28684 2.19052 2.23876 2.23861C2.19068 2.28669 2.16664 2.34813 2.16664 2.42292V7.37484Z" fill="#0275FF"/>
+  </svg>
+
+{:else if name == "page-links"}
+  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 17 18" fill="none">
+    <path d="M6 1.92773C3.08702 2.95733 1 5.73542 1 9.00098C1 13.1431 4.35787 16.501 8.5 16.501C12.6422 16.501 16 13.1431 16 9.00098C16 5.73542 13.913 2.95733 11 1.92773" stroke="#0275FF" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+  </svg>
+
+{:else if name == "sidebar"}
+  <svg xmlns="http://www.w3.org/2000/svg" width="30" height="30" viewBox="0 0 24 24" fill="#0275FF">
+    <path fill-rule="evenodd" d="M11.28 9.53L8.81 12l2.47 2.47a.75.75 0 11-1.06 1.06l-3-3a.75.75 0 010-1.06l3-3a.75.75 0 111.06 1.06z"/><path fill-rule="evenodd" d="M3.75 2A1.75 1.75 0 002 3.75v16.5c0 .966.784 1.75 1.75 1.75h16.5A1.75 1.75 0 0022 20.25V3.75A1.75 1.75 0 0020.25 2H3.75zM3.5 3.75a.25.25 0 01.25-.25H15v17H3.75a.25.25 0 01-.25-.25V3.75zm13 16.75v-17h3.75a.25.25 0 01.25.25v16.5a.25.25 0 01-.25.25H16.5z"/>
+  </svg>
+
+{:else if name == "accessdash-logo"}
+  <svg width="222" height="44" viewBox="0 0 145 29" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <title>AccessDash Logo</title>
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M4.19918 28.4917C5.31928 29.3332 6.90876 29.1064 7.74937 27.9851L10.7108 24.0351C11.1215 23.4872 11.3433 23.1913 11.3704 23.1544C13.3301 20.488 13.9654 17.3728 13.0587 14.528C12.152 11.6832 9.83167 9.51196 6.69148 8.47438C6.64806 8.46004 6.13772 8.29679 5.17006 7.98726L4.93328 7.91153C4.41679 7.74632 3.8929 7.57245 3.36388 7.38964C2.04009 6.93218 0.596484 7.63559 0.139495 8.96075C-0.317494 10.2859 0.385187 11.731 1.70898 12.1885C2.27663 12.3846 2.83766 12.5708 3.38961 12.7474L3.4102 12.754C3.96789 12.9323 4.3868 13.0663 4.671 13.1573C4.81489 13.2033 4.92389 13.2382 4.99876 13.2622C5.03995 13.2754 5.06693 13.2841 5.08307 13.2893C5.10015 13.2948 5.10508 13.2963 5.10186 13.2953C6.938 13.902 7.88602 15.0007 8.2272 16.0711C8.56839 17.1416 8.43121 18.5869 7.28537 20.146C7.28684 20.144 7.28606 20.145 7.28193 20.1505C7.27615 20.1583 7.2638 20.1748 7.24186 20.2041C7.20939 20.2475 7.16262 20.3099 7.10104 20.3921C6.97913 20.5548 6.80039 20.7932 6.56218 21.1109L3.69313 24.9378C2.85251 26.0591 3.07908 27.6502 4.19918 28.4917Z" fill="url(#paint0_linear_20_45)"/>
+    <path d="M15.5361 3.625C15.5361 5.62703 13.9132 7.25 11.9111 7.25C9.9091 7.25 8.28613 5.62703 8.28613 3.625C8.28613 1.62297 9.9091 0 11.9111 0C13.9132 0 15.5361 1.62297 15.5361 3.625Z" fill="url(#paint1_linear_20_45)"/>
+    <g style="mix-blend-mode:multiply">
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M19.6235 28.4949C18.4994 29.334 16.9073 29.1039 16.0674 27.9809L13.1025 24.0165C12.6922 23.4679 12.4706 23.1717 12.4436 23.1348C10.4897 20.4698 9.85591 17.3609 10.7639 14.521C11.6719 11.681 13.9923 9.51483 17.1307 8.47495C17.1741 8.46057 17.6863 8.29625 18.6576 7.98463L18.8937 7.90886C19.4052 7.74478 19.9238 7.57218 20.4475 7.39076C21.7733 6.93151 23.2207 7.63296 23.6804 8.9575C24.1401 10.282 23.438 11.7281 22.1122 12.1873C21.5499 12.3821 20.9942 12.567 20.4474 12.7425C19.8778 12.9252 19.4508 13.0622 19.1623 13.1548C19.018 13.2011 18.9087 13.2362 18.8336 13.2603C18.7486 13.2877 18.724 13.2957 18.7302 13.2936C16.8927 13.9025 15.9455 14.9989 15.6044 16.0657C15.2633 17.1325 15.3989 18.5745 16.5429 20.1349C16.5412 20.1325 16.5426 20.1344 16.5489 20.1428C16.5554 20.1515 16.5672 20.1674 16.5864 20.1931C16.6188 20.2365 16.6655 20.299 16.727 20.3812C16.8499 20.5456 17.0307 20.7874 17.272 21.1101L20.138 24.9421C20.9779 26.0651 20.7476 27.6558 19.6235 28.4949Z" fill="url(#paint2_linear_20_45)"/>
+    </g>
+    <path d="M36.3 23L41.9 8.88H44.74L50.28 23H46.96L42.72 11.28H43.86L39.56 23H36.3ZM39.48 20.44V17.88H47.18V20.44H39.48ZM55.862 23.22C54.862 23.22 53.9554 23 53.142 22.56C52.3287 22.12 51.6887 21.5133 51.222 20.74C50.7554 19.9667 50.522 19.1 50.522 18.14C50.522 17.1667 50.7554 16.3 51.222 15.54C51.702 14.7667 52.3487 14.16 53.162 13.72C53.9754 13.28 54.8887 13.06 55.902 13.06C56.662 13.06 57.3554 13.1933 57.982 13.46C58.622 13.7133 59.1887 14.1 59.682 14.62L57.722 16.58C57.4954 16.3267 57.2287 16.14 56.922 16.02C56.6287 15.9 56.2887 15.84 55.902 15.84C55.462 15.84 55.0687 15.94 54.722 16.14C54.3887 16.3267 54.122 16.5933 53.922 16.94C53.7354 17.2733 53.642 17.6667 53.642 18.12C53.642 18.5733 53.7354 18.9733 53.922 19.32C54.122 19.6667 54.3954 19.94 54.742 20.14C55.0887 20.34 55.4754 20.44 55.902 20.44C56.302 20.44 56.6554 20.3733 56.962 20.24C57.282 20.0933 57.5554 19.8933 57.782 19.64L59.722 21.6C59.2154 22.1333 58.642 22.54 58.002 22.82C57.362 23.0867 56.6487 23.22 55.862 23.22ZM65.5495 23.22C64.5495 23.22 63.6429 23 62.8295 22.56C62.0162 22.12 61.3762 21.5133 60.9095 20.74C60.4429 19.9667 60.2095 19.1 60.2095 18.14C60.2095 17.1667 60.4429 16.3 60.9095 15.54C61.3895 14.7667 62.0362 14.16 62.8495 13.72C63.6629 13.28 64.5762 13.06 65.5895 13.06C66.3495 13.06 67.0429 13.1933 67.6695 13.46C68.3095 13.7133 68.8762 14.1 69.3695 14.62L67.4095 16.58C67.1829 16.3267 66.9162 16.14 66.6095 16.02C66.3162 15.9 65.9762 15.84 65.5895 15.84C65.1495 15.84 64.7562 15.94 64.4095 16.14C64.0762 16.3267 63.8095 16.5933 63.6095 16.94C63.4229 17.2733 63.3295 17.6667 63.3295 18.12C63.3295 18.5733 63.4229 18.9733 63.6095 19.32C63.8095 19.6667 64.0829 19.94 64.4295 20.14C64.7762 20.34 65.1629 20.44 65.5895 20.44C65.9895 20.44 66.3429 20.3733 66.6495 20.24C66.9695 20.0933 67.2429 19.8933 67.4695 19.64L69.4095 21.6C68.9029 22.1333 68.3295 22.54 67.6895 22.82C67.0495 23.0867 66.3362 23.22 65.5495 23.22ZM75.337 23.22C74.2837 23.22 73.3437 23.0067 72.517 22.58C71.7037 22.14 71.0637 21.5333 70.597 20.76C70.1304 19.9867 69.897 19.1133 69.897 18.14C69.897 17.1667 70.1237 16.3 70.577 15.54C71.0437 14.7667 71.6704 14.16 72.457 13.72C73.2437 13.28 74.1304 13.06 75.117 13.06C76.077 13.06 76.9237 13.2667 77.657 13.68C78.3904 14.0933 78.9637 14.6667 79.377 15.4C79.8037 16.1333 80.017 16.9733 80.017 17.92C80.017 18.0933 80.0037 18.28 79.977 18.48C79.9637 18.6667 79.9304 18.8867 79.877 19.14L71.497 19.16V17.06L78.577 17.04L77.257 17.92C77.2437 17.36 77.157 16.9 76.997 16.54C76.837 16.1667 76.597 15.8867 76.277 15.7C75.9704 15.5 75.5904 15.4 75.137 15.4C74.657 15.4 74.237 15.5133 73.877 15.74C73.5304 15.9533 73.257 16.26 73.057 16.66C72.8704 17.06 72.777 17.5467 72.777 18.12C72.777 18.6933 72.877 19.1867 73.077 19.6C73.2904 20 73.5837 20.3133 73.957 20.54C74.3437 20.7533 74.797 20.86 75.317 20.86C75.797 20.86 76.2304 20.78 76.617 20.62C77.0037 20.4467 77.3437 20.1933 77.637 19.86L79.317 21.54C78.837 22.1 78.257 22.52 77.577 22.8C76.897 23.08 76.1504 23.22 75.337 23.22ZM85.0341 23.24C84.4607 23.24 83.8941 23.1667 83.3341 23.02C82.7874 22.8733 82.2741 22.6667 81.7941 22.4C81.3274 22.12 80.9274 21.8 80.5941 21.44L82.3341 19.68C82.6541 20.0267 83.0341 20.3 83.4741 20.5C83.9141 20.6867 84.3941 20.78 84.9141 20.78C85.2741 20.78 85.5474 20.7267 85.7341 20.62C85.9341 20.5133 86.0341 20.3667 86.0341 20.18C86.0341 19.94 85.9141 19.76 85.6741 19.64C85.4474 19.5067 85.1541 19.3933 84.7941 19.3C84.4341 19.1933 84.0541 19.08 83.6541 18.96C83.2541 18.84 82.8741 18.6733 82.5141 18.46C82.1541 18.2467 81.8607 17.9533 81.6341 17.58C81.4074 17.1933 81.2941 16.7067 81.2941 16.12C81.2941 15.4933 81.4541 14.9533 81.7741 14.5C82.0941 14.0333 82.5474 13.6667 83.1341 13.4C83.7207 13.1333 84.4074 13 85.1941 13C86.0207 13 86.7807 13.1467 87.4741 13.44C88.1807 13.72 88.7541 14.14 89.1941 14.7L87.4541 16.46C87.1474 16.1 86.8007 15.8467 86.4141 15.7C86.0407 15.5533 85.6741 15.48 85.3141 15.48C84.9674 15.48 84.7074 15.5333 84.5341 15.64C84.3607 15.7333 84.2741 15.8733 84.2741 16.06C84.2741 16.26 84.3874 16.42 84.6141 16.54C84.8407 16.66 85.1341 16.7667 85.4941 16.86C85.8541 16.9533 86.2341 17.0667 86.6341 17.2C87.0341 17.3333 87.4141 17.5133 87.7741 17.74C88.1341 17.9667 88.4274 18.2733 88.6541 18.66C88.8807 19.0333 88.9941 19.5267 88.9941 20.14C88.9941 21.0867 88.6341 21.84 87.9141 22.4C87.2074 22.96 86.2474 23.24 85.0341 23.24ZM94.1942 23.24C93.6209 23.24 93.0542 23.1667 92.4942 23.02C91.9476 22.8733 91.4342 22.6667 90.9542 22.4C90.4876 22.12 90.0876 21.8 89.7542 21.44L91.4942 19.68C91.8142 20.0267 92.1942 20.3 92.6342 20.5C93.0742 20.6867 93.5542 20.78 94.0742 20.78C94.4342 20.78 94.7076 20.7267 94.8942 20.62C95.0942 20.5133 95.1942 20.3667 95.1942 20.18C95.1942 19.94 95.0742 19.76 94.8342 19.64C94.6076 19.5067 94.3142 19.3933 93.9542 19.3C93.5942 19.1933 93.2142 19.08 92.8142 18.96C92.4142 18.84 92.0342 18.6733 91.6742 18.46C91.3142 18.2467 91.0209 17.9533 90.7942 17.58C90.5676 17.1933 90.4542 16.7067 90.4542 16.12C90.4542 15.4933 90.6142 14.9533 90.9342 14.5C91.2542 14.0333 91.7076 13.6667 92.2942 13.4C92.8809 13.1333 93.5676 13 94.3542 13C95.1809 13 95.9409 13.1467 96.6342 13.44C97.3409 13.72 97.9142 14.14 98.3542 14.7L96.6142 16.46C96.3076 16.1 95.9609 15.8467 95.5742 15.7C95.2009 15.5533 94.8342 15.48 94.4742 15.48C94.1276 15.48 93.8676 15.5333 93.6942 15.64C93.5209 15.7333 93.4342 15.8733 93.4342 16.06C93.4342 16.26 93.5476 16.42 93.7742 16.54C94.0009 16.66 94.2942 16.7667 94.6542 16.86C95.0142 16.9533 95.3942 17.0667 95.7942 17.2C96.1942 17.3333 96.5742 17.5133 96.9342 17.74C97.2942 17.9667 97.5876 18.2733 97.8142 18.66C98.0409 19.0333 98.1542 19.5267 98.1542 20.14C98.1542 21.0867 97.7942 21.84 97.0742 22.4C96.3676 22.96 95.4076 23.24 94.1942 23.24ZM101.614 23V21.28H105.474C106.501 21.28 107.401 21.06 108.174 20.62C108.948 20.1667 109.548 19.5467 109.974 18.76C110.414 17.9733 110.634 17.0667 110.634 16.04C110.634 15.0267 110.414 14.1267 109.974 13.34C109.534 12.5533 108.928 11.94 108.154 11.5C107.381 11.06 106.488 10.84 105.474 10.84H101.634V9.12H105.514C106.528 9.12 107.461 9.29333 108.314 9.64C109.181 9.98667 109.928 10.4733 110.554 11.1C111.194 11.7267 111.688 12.46 112.034 13.3C112.394 14.14 112.574 15.06 112.574 16.06C112.574 17.0467 112.394 17.9667 112.034 18.82C111.688 19.66 111.201 20.3933 110.574 21.02C109.948 21.6467 109.201 22.1333 108.334 22.48C107.481 22.8267 106.554 23 105.554 23H101.614ZM100.374 23V9.12H102.254V23H100.374ZM118.899 23.2C118.032 23.2 117.246 22.9867 116.539 22.56C115.832 22.12 115.272 21.5267 114.859 20.78C114.459 20.0333 114.259 19.1933 114.259 18.26C114.259 17.3267 114.459 16.4867 114.859 15.74C115.272 14.9933 115.826 14.4 116.519 13.96C117.226 13.52 118.019 13.3 118.899 13.3C119.619 13.3 120.259 13.4533 120.819 13.76C121.392 14.0533 121.852 14.4667 122.199 15C122.546 15.52 122.739 16.1267 122.779 16.82V19.68C122.739 20.36 122.546 20.9667 122.199 21.5C121.866 22.0333 121.412 22.4533 120.839 22.76C120.279 23.0533 119.632 23.2 118.899 23.2ZM119.199 21.5C120.092 21.5 120.812 21.2 121.359 20.6C121.906 19.9867 122.179 19.2067 122.179 18.26C122.179 17.6067 122.052 17.04 121.799 16.56C121.559 16.0667 121.212 15.6867 120.759 15.42C120.306 15.14 119.779 15 119.179 15C118.579 15 118.046 15.14 117.579 15.42C117.126 15.7 116.766 16.0867 116.499 16.58C116.246 17.06 116.119 17.6133 116.119 18.24C116.119 18.88 116.246 19.4467 116.499 19.94C116.766 20.42 117.132 20.8 117.599 21.08C118.066 21.36 118.599 21.5 119.199 21.5ZM122.059 23V20.44L122.399 18.12L122.059 15.82V13.5H123.879V23H122.059ZM129.52 23.2C128.986 23.2 128.48 23.1333 128 23C127.533 22.8533 127.1 22.6533 126.7 22.4C126.3 22.1333 125.953 21.82 125.66 21.46L126.82 20.3C127.166 20.7267 127.566 21.0467 128.02 21.26C128.473 21.46 128.98 21.56 129.54 21.56C130.1 21.56 130.533 21.4667 130.84 21.28C131.146 21.08 131.3 20.8067 131.3 20.46C131.3 20.1133 131.173 19.8467 130.92 19.66C130.68 19.46 130.366 19.3 129.98 19.18C129.593 19.0467 129.18 18.92 128.74 18.8C128.313 18.6667 127.906 18.5 127.52 18.3C127.133 18.1 126.813 17.8267 126.56 17.48C126.32 17.1333 126.2 16.6733 126.2 16.1C126.2 15.5267 126.34 15.0333 126.62 14.62C126.9 14.1933 127.286 13.8667 127.78 13.64C128.286 13.4133 128.893 13.3 129.6 13.3C130.346 13.3 131.006 13.4333 131.58 13.7C132.166 13.9533 132.646 14.34 133.02 14.86L131.86 16.02C131.593 15.6733 131.26 15.4067 130.86 15.22C130.473 15.0333 130.033 14.94 129.54 14.94C129.02 14.94 128.62 15.0333 128.34 15.22C128.073 15.3933 127.94 15.64 127.94 15.96C127.94 16.28 128.06 16.5267 128.3 16.7C128.54 16.8733 128.853 17.02 129.24 17.14C129.64 17.26 130.053 17.3867 130.48 17.52C130.906 17.64 131.313 17.8067 131.7 18.02C132.086 18.2333 132.4 18.52 132.64 18.88C132.893 19.24 133.02 19.7133 133.02 20.3C133.02 21.1933 132.7 21.9 132.06 22.42C131.433 22.94 130.586 23.2 129.52 23.2ZM141.755 23V17.46C141.755 16.74 141.528 16.1467 141.075 15.68C140.622 15.2133 140.035 14.98 139.315 14.98C138.835 14.98 138.408 15.0867 138.035 15.3C137.662 15.5133 137.368 15.8067 137.155 16.18C136.942 16.5533 136.835 16.98 136.835 17.46L136.095 17.04C136.095 16.32 136.255 15.68 136.575 15.12C136.895 14.56 137.342 14.12 137.915 13.8C138.488 13.4667 139.135 13.3 139.855 13.3C140.575 13.3 141.208 13.46 141.755 13.78C142.315 14.1 142.755 14.5467 143.075 15.12C143.395 15.6933 143.555 16.3533 143.555 17.1V23H141.755ZM135.035 23V8.72H136.835V23H135.035Z" fill="#0275FF"/>
+    <defs>
+    <linearGradient id="paint0_linear_20_45" x1="6.12576" y1="26.3964" x2="-2.9966" y2="22.3913" gradientUnits="userSpaceOnUse">
+    <stop stop-color="#17ACFF"/>
+    <stop offset="1" stop-color="#4DD78A"/>
+    </linearGradient>
+    <linearGradient id="paint1_linear_20_45" x1="11.9111" y1="0" x2="11.9111" y2="7.25" gradientUnits="userSpaceOnUse">
+    <stop stop-color="#1EB2F1"/>
+    <stop offset="1" stop-color="#4ED88B"/>
+    </linearGradient>
+    <linearGradient id="paint2_linear_20_45" x1="17.6992" y1="26.3964" x2="26.8066" y2="22.4074" gradientUnits="userSpaceOnUse">
+    <stop stop-color="#17ACFF"/>
+    <stop offset="1" stop-color="#4DD78A"/>
+    </linearGradient>
+    </defs>
+  </svg>
+
+
+  {:else if name == "good"}
+  <svg class="top-icon" width="20" height="20" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1000 1000">
+    <path fill="#118c0b" d="M500,131c-203.8,0-369,165.2-369,369s165.2,369,369,369,369-165.21,369-369-165.21-369-369-369ZM437.92,707.75l-185.29-185.29,70.71-70.71,114.58,114.58,238.73-238.73,70.71,70.71-309.44,309.44Z"/>
+  </svg>
+  
+  
+  {:else if name == "warning"}
+  <svg class="top-icon" width="20" height="20" viewBox="0 0 512 512" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" >
+    <path d="M246.312928,5.62892705 C252.927596,9.40873724 258.409564,14.8907053 262.189374,21.5053731 L444.667042,340.84129 C456.358134,361.300701 449.250007,387.363834 428.790595,399.054926 C422.34376,402.738832 415.04715,404.676552 407.622001,404.676552 L42.6666667,404.676552 C19.1025173,404.676552 7.10542736e-15,385.574034 7.10542736e-15,362.009885 C7.10542736e-15,354.584736 1.93772021,347.288125 5.62162594,340.84129 L188.099293,21.5053731 C199.790385,1.04596203 225.853517,-6.06216498 246.312928,5.62892705 Z M224,272 C208.761905,272 197.333333,283.264 197.333333,298.282667 C197.333333,313.984 208.415584,325.248 224,325.248 C239.238095,325.248 250.666667,313.984 250.666667,298.624 C250.666667,283.264 239.238095,272 224,272 Z M245.333333,106.666667 L202.666667,106.666667 L202.666667,234.666667 L245.333333,234.666667 L245.333333,106.666667 Z" fill="#FF8D02"></path>
+  </svg>
+
+  
+  
+  {:else if name == "critical-warning"}
+  <svg class="top-icon" width="20" height="20" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M19.5099 5.85L13.5699 2.42C12.5999 1.86 11.3999 1.86 10.4199 2.42L4.48992 5.85C3.51992 6.41 2.91992 7.45 2.91992 8.58V15.42C2.91992 16.54 3.51992 17.58 4.48992 18.15L10.4299 21.58C11.3999 22.14 12.5999 22.14 13.5799 21.58L19.5199 18.15C20.4899 17.59 21.0899 16.55 21.0899 15.42V8.58C21.0799 7.45 20.4799 6.42 19.5099 5.85ZM11.2499 7.75C11.2499 7.34 11.5899 7 11.9999 7C12.4099 7 12.7499 7.34 12.7499 7.75V13C12.7499 13.41 12.4099 13.75 11.9999 13.75C11.5899 13.75 11.2499 13.41 11.2499 13V7.75ZM12.9199 16.63C12.8699 16.75 12.7999 16.86 12.7099 16.96C12.5199 17.15 12.2699 17.25 11.9999 17.25C11.8699 17.25 11.7399 17.22 11.6199 17.17C11.4899 17.12 11.3899 17.05 11.2899 16.96C11.1999 16.86 11.1299 16.75 11.0699 16.63C11.0199 16.51 10.9999 16.38 10.9999 16.25C10.9999 15.99 11.0999 15.73 11.2899 15.54C11.3899 15.45 11.4899 15.38 11.6199 15.33C11.9899 15.17 12.4299 15.26 12.7099 15.54C12.7999 15.64 12.8699 15.74 12.9199 15.87C12.9699 15.99 12.9999 16.12 12.9999 16.25C12.9999 16.38 12.9699 16.51 12.9199 16.63Z" fill="#FF0202"/>
+    </svg>
+{/if}
+
+<style>
+
+svg {
+  min-width: 20px;
+}
+
+</style>
