+++ src/lib/index.js
@@ -1,19 +1,16 @@
-// place files you want to import through the `$lib` alias in this folder.
-
-export { default as Icon } from './1_Atoms/Icon.svelte'
-export { default as BlankBarChart } from "./ReusableComponents/BlankBarChart.svelte";
-export { default as SkipLink } from './1_Atoms/SkipLink.svelte'
-
-export { default as BarChart } from './2_Molecules/BarChart.svelte'
-export { default as DonutChart } from './2_Molecules/DonutChart.svelte'
-export { default as MonthlyNavigation } from './2_Molecules/MonthlyNavigation.svelte'
-
-export { default as Title } from './3_Organism/Title.svelte'
-export { default as AccessibillityGraph } from './3_Organism/AccessibillityGraph.svelte'
-export { default as AutomaticScan } from "./Pages/Home/AutomaticScan.svelte";
-export { default as ErrorResult } from './3_Organism/ErrorResult.svelte'
+/*REUSABLE COMPONENTS*/
+export { default as Header } from "./ReusableComponents/Header.svelte";
+export { default as Footer } from "./ReusableComponents/Footer.svelte";
+export { default as SkipLink } from "./ReusableComponents/SkipLink.svelte";
+export { default as Icon } from "./ReusableComponents/Icon.svelte";
 export { default as BlankDonutChart } from "./ReusableComponents/BlankDonutChart.svelte";
-export { default as TypeGraph } from "./Pages/Home/TypeGraph.svelte";
-export { default as Header } from './3_Organism/Header.svelte'
-export { default as Footer } from './3_Organism/Footer.svelte'
+export { default as BlankLineChart } from "./ReusableComponents/BlankLineChart.svelte";
+export { default as BlankBarChart } from "./ReusableComponents/BlankBarChart.svelte";
 
+/*HOME*/
+export { default as Title } from "./Pages/Home/Title.svelte";
+export { default as AccessibillityGraph } from "./Pages/Home/AccessibillityGraph.svelte";
+export { default as AutomaticScan } from "./Pages/Home/AutomaticScan.svelte";
+export { default as ErrorResult } from "./Pages/Home/ErrorResult.svelte";
+export { default as TypeGraph } from "./Pages/Home/TypeGraph.svelte";
+export { default as MonthlyNavigation } from "./Pages/Home/MonthlyNavigation.svelte";
